Allow tagd to force login route to pre/post 1.7.0
Some vendors like OVH either omit the version information or run
a custom build that returns a non-semver version tag. tagd should
still work with these deployments of harbor, so it learned a new
argument:

--login-behavior <Probe|ForcePre17|ForcePost17>

Probe: The current default behavior, use the /api/systeminfo
endpoint to detect the version of harbor.

ForcePre17: Assume the version of harbor is older than 1.7

ForcePost17: Assume the version of harbor is 1.7 or newer